fix relinking of /bin/sh in Conf.d/build-env/bin_bash.sh

+++ source/Conf.d/build-env/bin_bash.sh
@@ -7,7 +7,7 @@ check_required_files $CPREFIX/bin/bash
 
 # relink /bin/sh -> $bash, keep original /bin/sh as /bin/sh.orig
 if [ "$(basename $(readlink -f /bin/sh))" != "bash" ]; then
-	mv /bin/sh /bin/sh.orig
-	ln -s $bash /bin/sh
+	[ -e /bin/sh ] && mv /bin/sh /bin/sh.orig
+	ln -s $CPREFIX/bin/bash /bin/sh
 fi
